refactor(api): extract project activity projection
