+++ docs-ai/skills/backend/SKILL.md
@@ -126,6 +126,11 @@ runner orchestration to `handler_project_work.go`. Keep app-layer dependencies
 narrow: define the minimal store/runner interfaces the command needs instead
 of accepting broad subsystem stores by habit.
 
+Project activity is a read/projection surface: keep its response DTOs,
+linked-chat summaries, artifact/handoff signal grouping, and bucket/status
+summaries in `internal/api/handler_project_activity.go`. Avoid adding activity
+projection helpers back to `handler_project_work.go`.
+
 ### Change chat-session / ACP adapter behavior
 
 Chat sessions separate ownership from turn execution:
